ode-217 refactored J2735TravelerInputData
