+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/pdm/J2735ProbeDataManagementTest.java
@@ -8,7 +8,7 @@ import org.junit.runner.RunWith;
 import mockit.Tested;
 import mockit.integration.junit4.JMockit;
 import us.dot.its.jpo.ode.plugin.RoadSideUnit.RSU;
-import us.dot.its.jpo.ode.plugin.j2735.J2735TravelerInputData.ODE;
+import us.dot.its.jpo.ode.plugin.OperationalDataEnvironment.ODE;
 
 @RunWith(JMockit.class)
 public class J2735ProbeDataManagementTest {
